+++ LoopStructural/modelling/fault/fault_builder.py
@@ -13,7 +13,7 @@ class FaultBuilder(StructuralFrameBuilder):
         interpolators=None,
         model=None,
         fault_bounding_box_buffer=0.2,
-        **kwargs
+        **kwargs,
     ):
         """A specialised structural frame builder for building a fault
 
@@ -31,9 +31,9 @@ class FaultBuilder(StructuralFrameBuilder):
         """
 
         StructuralFrameBuilder.__init__(self, interpolator, interpolators, **kwargs)
-        self.origin = np.array([np.nan, np.nan, np.nan])
-        self.maximum = np.array([np.nan, np.nan, np.nan])
         self.model = model
+        self.origin = np.zeros(3)  # self.model.bounding_box[0, :]
+        self.maximum = np.zeros(3)  # self.model.bounding_box[1, :]
         # define a maximum area to mesh adding buffer to model
         # buffer = .2
         self.minimum_origin = self.model.bounding_box[
@@ -100,11 +100,15 @@ class FaultBuilder(StructuralFrameBuilder):
             distance = np.linalg.norm(
                 fault_trace[:, None, :] - fault_trace[None, :, :], axis=2
             )
-            if len(distance) == 0:
+            if len(distance) == 0 or np.sum(distance) == 0:
                 logger.error("There is no fault trace for {}".format(self.name))
+                # add any data anyway - usually just orientation data
+                self.add_data_from_data_frame(data)
+                self.origin = self.model.bounding_box[0, :]
+                self.maximum = self.model.bounding_box[1, :]
                 return
             major_axis = np.max(distance)
-            logger.warning("Fault major axis using map length: {}".format(major_axis))
+            logger.warning(f"Fault major axis using map length: {major_axis}")
 
         if minor_axis is None:
             minor_axis = major_axis / 2.0
@@ -213,13 +217,10 @@ class FaultBuilder(StructuralFrameBuilder):
                     1,
                     w,
                 ]
-                # data.loc[len(data),['X','Y','Z','feature_name','val','coord']] = \
-                #     [fault_depth[1,0],fault_depth[1,1],fault_depth[1,2],self.name,-1,1]
+
                 self.update_geometry(fault_depth)
                 # TODO need to add data here
-                # print(np.linalg.norm(slip_vector))
-                # slip_vector /= intermediate_axis
-                # print(np.linalg.norm(slip_vector))
+
                 data.loc[
                     len(data),
                     [
@@ -246,9 +247,6 @@ class FaultBuilder(StructuralFrameBuilder):
                     1,
                     w,
                 ]
-        # add strike vector to constraint fault extent
-        # data.loc[len(data),['X','Y','Z','feature_name','nx','ny','nz','coord']] = [fault_center[0],fault_center[1],fault_center[2],\
-        #     self.name, strike_vector[0], strike_vector[1], strike_vector[2], 2]
         self.add_data_from_data_frame(data)
         self.update_geometry(data[["X", "Y", "Z"]].to_numpy())
 
@@ -261,13 +259,6 @@ class FaultBuilder(StructuralFrameBuilder):
             percentage of length to add to edges
         """
         length = np.max(self.maximum - self.origin)
-        # origin = self.builders[0].interpolator.support.origin
-        # maximum = self.builders[0].interpolator.support.maximum#set_interpolation_geometry
-        # if origin[2]>self.origin[2]:
-        #     origin[2]=self.origin[2]
-        # if maximum[2]<self.maximum[2]:
-        #     maximum[2]=self.maximum[2]
-        # self.builders[0].set_interpolation_geometry(origin,maximum)
         # for builder in self.builders:
         # all three coordinates share the same support
         self.builders[0].set_interpolation_geometry(
@@ -298,9 +289,7 @@ class FaultBuilder(StructuralFrameBuilder):
                     return mask
                 else:
                     logger.warning(
-                        "Not adding splay, cannot identify splay overlap region for {} and {}".format(
-                            self.name, splay.name
-                        )
+                        f"Not adding splay, cannot identify splay overlap region for {self.name} and {splay.name}"
                     )
                     return mask